Code review fixes, made seperate helper file
